@@ -8,9 +8,11 @@ import (
 	"github.com/Khan/genqlient/graphql"
 	"github.com/aquasecurity/trivy/pkg/commands/artifact"
 	"github.com/aquasecurity/trivy/pkg/flag"
+	aqualog "github.com/aquasecurity/trivy/pkg/log"
 	"github.com/aquasecurity/trivy/pkg/types"
 	"github.com/uselagoon/lagoon/services/insights-handler/internal/lagoonclient"
 	"io"
+	"log/slog"
 	"net/http"
 	"os"
 	"strings"
@@ -83,7 +85,12 @@ func writeProblemsArrayToApi(apiClient graphql.Client, environment int, source s
 	if err != nil {
 		return err
 	}
-	fmt.Printf("Deleted problems from API for %v:%v - response: %v\n", service, source, ret)
+	//fmt.Printf("Deleted problems from API for %v:%v - response: %v\n", service, source, ret)
+	slog.Info("Deleted problems from API",
+		"Service", service,
+		"Source", source,
+		"Return Data", ret,
+	)
 
 	//now we write the problems themselves
 	_, err = lagoonclient.AddProblems(context.TODO(), apiClient, problems)
@@ -145,7 +152,6 @@ func executeProcessingTrivy(trivyRemoteAddress string, bomWriteDir string, bom c
 		os.Remove(fullFilename)
 		file.Close()
 	}()
-
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*1000)
 	defer cancel()
 
@@ -239,7 +245,18 @@ func trivyReportToProblems(environment int, source string, service string, repor
 			ret = append(ret, p)
 		}
 	}
-	fmt.Printf("Found %v problems for environment %v\n", len(ret), environment)
+	//fmt.Printf("Found %v problems for environment %v\n", len(ret), environment)
+	slog.Info("Found problems",
+		"EnvironmentId", environment,
+		"Source", source,
+		"Number", len(ret),
+	)
 
 	return ret, nil
+}
+
+func init() {
+	// Ensure that logging is turned off for Zap so that our logs aren't smashed by Trivy.
+	// Only errors are used for output
+	aqualog.InitLogger(false, true)
 }
